moved popups to ui folder
