+++ src/core/mod.rs
@@ -2,7 +2,6 @@
 
 pub mod app;
 pub mod editor;
-pub mod popups;
 pub mod cursor;
 pub mod command_line;
 pub mod errors;
