+++ src/ui/mod.rs
@@ -1,1 +1,2 @@
 pub mod ui;
+pub mod popups;
